gadget: mv ensureLayoutCompatibility to gadget proper, add gadgettest pkg
Move ensureLayoutCompatibility and related helpers to the top-level gadget
package so that they can be used from the gadget asset update routine.

Also move some gadget test helpers that are now shared between multiple gadget
test packages (install_test and gadget_test now) to a shared gadgettest package
where we can easily share gadget helpers. This package can probably grow more
helpers but this is the bare minimum needed to move forward.
